requirements.txt
# Python package dependencies for PathoGen
# Core runtime dependencies and optional development/testing packages

# SYSTEM REQUIREMENTS:
# - perf tool (linux-perf package) - REQUIRED for instruction counting
#   Ubuntu/Debian: sudo apt-get install linux-perf
#   RHEL/CentOS: sudo yum install perf
#   Arch Linux: sudo pacman -S perf

# Core dependencies
PyYAML>=6.0
requests>=2.28.0
psutil>=5.9.0
pathlib  # Usually included in Python 3.4+, but explicit for clarity

# Reporting and visualization
matplotlib>=3.5.0
seaborn>=0.11.0
numpy>=1.21.0

# LLM providers for performance analysis
# OpenAI - reliable and powerful models with excellent algorithmic reasoning
openai>=1.0.0

# Groq - uses requests library (already included above)
# Ultra-fast inference with 70B parameter models, free tier available

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
isort>=5.10.0
flake8>=5.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Additional runtime dependencies
# Standard library modules (included for reference):
# - ast (built-in)
# - json (built-in)
# - re (built-in)  
# - os (built-in)
# - sys (built-in)
# - time (built-in)
# - signal (built-in)
# - subprocess (built-in)
# - tempfile (built-in)
# - logging (built-in)
# - argparse (built-in)
# - collections (built-in)
# - dataclasses (built-in Python 3.7+)
# - enum (built-in Python 3.4+)
# - statistics (built-in Python 3.4+)
# - typing (built-in Python 3.5+)
# - abc (built-in)
# - shutil (built-in)
# - traceback (built-in)
# - stat (built-in)
